requirements.txt
# Azure Functions runtime
azure-functions==1.21.3

# Azure Communication Services — envio de email
azure-communication-email==1.0.0

# Automação de browser
selenium==4.27.1
webdriver-manager==4.0.2

# Leitura de ficheiros Excel
pandas==3.0.0
openpyxl==3.1.5
xlrd==2.0.2
numpy==2.4.1

# Microsoft Graph API — autenticação Azure
msal==1.34.0
requests==2.32.5

# Utilitários de datas
python-dateutil==2.9.0.post0

# Variáveis de ambiente
python-dotenv==1.2.1
